pyproject.toml
[tool.black]
line-length = 95
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[tool.poetry]
name = "news-analytics"
version = "0.1.0"
description = ""
authors = ["Silvio Pavanetto <silvio.pavanetto@polimi.it>, Lorenzo Corti <lorenzo.corti@polimi.it, Andrea Tocchetti <andrea.tocchetti@polimi.it>"]

[tool.poetry.dependencies]
python = "3.7.6"
black = "^19.10b0"
gensim = "^3.8.3"
ipykernel = "^5.1.4"
ipython = "^7.13.0"
ipython-genutils = "^0.2.0"
json5 = "^0.9.0"
jsonschema = "^3.2.0"
jupyter-client = "^6.1.3"
jupyter-core = "^4.6.3"
jupyterlab = "^2.1.4"
jupyterlab-server = "^1.1.5"
nbformat = "^5.0.6"
notebook = "^6.0.3"
numpy = "^1.18.5"
pandas = "^1.0.4"
pre-commit = "^2.4.0"
pymongo = "^3.10.1"
PyYAML = "^5.3.1"
requests = "^2.23.0"
scipy = "^1.4.1"
toml = "^0.10.1"
urllib3 = "^1.25.8"
virtualenv = "^20.0.21"
networkx = "^2.4"
sklearn = "^0.0"
dateparser = "^0.7.5"
tensorflow = "^2.2.0"
spacy = "^2.2.4"
nltk = "^3.5"
bert-serving-server = "^1.10.0"
bert-serving-client = "^1.10.0"
neuralcoref = "^4.0"
transformers = "^2.11.0"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
en_core_web_md = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.2.5/en_core_web_md-2.2.5.tar.gz"}
de_core_news_md = {url = "https://github.com/explosion/spacy-models/releases/download/de_core_news_md-2.2.5/de_core_news_md-2.2.5.tar.gz"}
fr_core_news_md = {url = "https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-2.2.5/fr_core_news_md-2.2.5.tar.gz"}
it_core_news_sm = {url = "https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-2.2.5/it_core_news_sm-2.2.5.tar.gz"}
nl_core_news_sm = {url = "https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-2.2.5/nl_core_news_sm-2.2.5.tar.gz"}
es_core_news_md = {url = "https://github.com/explosion/spacy-models/releases/download/es_core_news_md-2.2.5/es_core_news_md-2.2.5.tar.gz"}
